JWT finished

+++ VacPlanner/apis/ConfigApi.js
@@ -57,6 +57,15 @@ module.exports = class ConfigApi {
       reservationField.add(ctx, next);
     });
     router.post("/assignmentCriteria", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "assignment_criteria_add",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       //Step 1 - Agregar a la bd y recuperar el id
       const id = await this.countryDataAccess.addCriteria(
         ctx.request.body.function
@@ -88,98 +97,90 @@ module.exports = class ConfigApi {
       ctx.status = 200;
       return;
     });
-
     router.post("/states", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "state_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       ctx.response.body = await stateController.addStates(ctx.request.body);
     });
     router.post("/zones", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "zone_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       await zoneController
         .addZones(ctx.request.body)
         .then((data) => (ctx.response.body = data))
         .catch((e) => (ctx.response.body = "stateCode no existe"));
     });
     router.post("/vaccenters", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "vac_center_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       await vacCenterController
         .addVacCenters(ctx.request.body)
         .then((data) => (ctx.response.body = data))
         .catch((e) => (ctx.response.body = "zoneId no existe"));
     });
     router.post("/vaccines", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "vaccine_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       await vaccineController
         .addVaccines(ctx.request.body)
         .then((data) => (ctx.response.body = data))
         .catch((e) => (ctx.response.body = "zoneId no existe"));
     });
     router.post("/vaccinationperiods", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "vac_period_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       await vaccinationPeriodController
         .addVaccinationPeriod(ctx.request.body)
         .then((data) => (ctx.response.body = data))
         .catch((e) => (ctx.response.body = e));
     });
-    router.post("/slots", async (ctx, next) => {
-      await slotController
-        .addSlot(ctx.request.body)
-        .then((data) => (ctx.response.body = data))
-        .catch((e) => (ctx.response.body = e));
-    });
-
-    //GET ALL
-    router.get("/states", async (ctx, next) => {
-      const res = await stateController.getStates();
-      ctx.response.body = res;
-    });
-    router.get("/zones", async (ctx, next) => {
-      const res = await zoneController.getZones();
-      ctx.response.body = res;
-    });
-    router.get("/vaccenters", async (ctx, next) => {
-      const res = await vacCenterController.getVacCenters();
-      ctx.response.body = res;
-    });
-    router.get("/vaccines", async (ctx, next) => {
-      const res = await vaccineController.getVaccines();
-      ctx.response.body = res;
-    });
-    router.get("/vaccinationperiods", async (ctx, next) => {
-      const res = await vaccinationPeriodController.getVaccinationPeriods();
-      ctx.response.body = res;
-    });
-    router.get("/slots/0", async (ctx, next) => {
-      console.log("here");
-      const res = await slotController.getSlots();
-      ctx.response.body = res;
-    });
-
-    //GET
-    router.get("/states/:code", async (ctx, next) => {
-      const res = await stateController.getAState(ctx.params.code);
-      ctx.response.body = res;
-    });
-    router.get("/zones/:id", async (ctx, next) => {
-      const res = await zoneController.getAZone(ctx.params.id);
-      ctx.response.body = res;
-    });
-    router.get("/vaccenters/:id", async (ctx, next) => {
-      const res = await vacCenterController.getAVacCenter(ctx.params.id);
-      ctx.response.body = res;
-    });
-    router.get("/vaccines/:id", async (ctx, next) => {
-      const res = await vaccineController.getAVaccine(ctx.params.id);
-      ctx.response.body = res;
-    });
-    router.get("/vaccinationperiods/:id", async (ctx, next) => {
-      const res = await vaccinationPeriodController.getAVaccinationPeriod(
-        ctx.params.id
-      );
-      ctx.response.body = res;
-    });
-    router.get("/slots", async (ctx, next) => {
-      const res = await slotController.getASlot(ctx.request.body);
-      ctx.response.body = res;
-    });
 
     //DELETE
     router.delete("/states/:code", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "state_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       const res = await stateController.deleteAState(ctx.params.code);
       if (res) {
         ctx.response.body = "Eliminada satisfactoriamente";
@@ -188,6 +189,15 @@ module.exports = class ConfigApi {
       }
     });
     router.delete("/zones/:id", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "zone_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       const res = await zoneController.deleteAZone(ctx.params.id);
       if (res) {
         ctx.response.body = "Eliminada satisfactoriamente";
@@ -196,6 +206,15 @@ module.exports = class ConfigApi {
       }
     });
     router.delete("/vaccenters/:id", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "vac_center_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       const res = await vacCenterController.deleteAVacCenter(ctx.params.id);
       if (res) {
         ctx.response.body = "Eliminada satisfactoriamente";
@@ -204,6 +223,15 @@ module.exports = class ConfigApi {
       }
     });
     router.delete("/vaccines/:id", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "vaccine_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       const res = await vaccineController.deleteAVaccine(ctx.params.id);
       if (res) {
         ctx.response.body = "Eliminada satisfactoriamente";
@@ -212,6 +240,15 @@ module.exports = class ConfigApi {
       }
     });
     router.delete("/vaccinationperiods/:id", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "vac_period_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       const res = await vaccinationPeriodController.deleteAVaccinationPeriod(
         ctx.params.id
       );
@@ -221,17 +258,18 @@ module.exports = class ConfigApi {
         ctx.response.body = "No se pudo eliminar, el codigo no existe";
       }
     });
-    router.delete("/slots", async (ctx, next) => {
-      const res = await slotController.deleteASlot(ctx.request.body);
-      if (res) {
-        ctx.response.body = "Eliminada satisfactoriamente";
-      } else {
-        ctx.response.body = "No se pudo eliminar, el codigo no existe";
-      }
-    });
 
     //UPDATE
     router.put("/states/:code", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "state_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       const res = await stateController.updateAState(
         ctx.params.code,
         ctx.request.body
@@ -243,6 +281,15 @@ module.exports = class ConfigApi {
       }
     });
     router.put("/zones/:id", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "zone_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       const res = await zoneController.updateAZone(
         ctx.params.id,
         ctx.request.body
@@ -254,6 +301,15 @@ module.exports = class ConfigApi {
       }
     });
     router.put("/vaccenters/:id", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "vac_center_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       const res = await vacCenterController.updateAVacCenter(
         ctx.params.id,
         ctx.request.body
@@ -265,6 +321,15 @@ module.exports = class ConfigApi {
       }
     });
     router.put("/vaccines/:id", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "vaccine_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       const res = await vaccineController.updateAVaccine(
         ctx.params.id,
         ctx.request.body
@@ -276,18 +341,19 @@ module.exports = class ConfigApi {
       }
     });
     router.put("/vaccinationperiods/:id", async (ctx, next) => {
+      const token = ctx.request.headers["authorization"].split("Bearer ")[1];
+      const hasPermission = await authController.checkPermissions(token, [
+        "vac_period_crud",
+      ]);
+      if (!hasPermission) {
+        ctx.response.body = "Unauthorized";
+        ctx.response.status = 401;
+        return;
+      }
       const res = await vaccinationPeriodController.updateAVaccinationPeriod(
         ctx.params.id,
         ctx.request.body
       );
-      if (res) {
-        ctx.response.body = "Modificado correctamente";
-      } else {
-        ctx.response.body = "No se pudo modificar, el codigo no existe";
-      }
-    });
-    router.put("/slots", async (ctx, next) => {
-      const res = await slotController.updateASlot(ctx.request.body);
       if (res) {
         ctx.response.body = "Modificado correctamente";
       } else {
